refactor reward related structs and methods
