+++ external_parser/example_joiner.cc
@@ -1,6 +1,7 @@
 #include "example_joiner.h"
 #include "log_converter.h"
 
+#include "generated/v2/CbEvent_generated.h"
 #include "generated/v2/DedupInfo_generated.h"
 #include "generated/v2/Event_generated.h"
 #include "generated/v2/Metadata_generated.h"
@@ -11,90 +12,16 @@
 #include <time.h>
 
 // VW headers
-#include "example.h"
 #include "io/logger.h"
 #include "parse_example_json.h"
 #include "parser.h"
-#include "v_array.h"
-
-namespace RewardFunctions {
-float average(const joined_event &event) {
-  float sum = 0.f;
-  for (const auto &o : event.outcome_events) {
-    sum += o.value;
-  }
-
-  return sum / event.outcome_events.size();
-}
-
-float sum(const joined_event &event) {
-  float sum = 0.f;
-  for (const auto &o : event.outcome_events) {
-    sum += o.value;
-  }
-
-  return sum;
-}
-
-float min(const joined_event &event) {
-  float min_reward = std::numeric_limits<float>::max();
-  for (const auto &o : event.outcome_events) {
-    if (o.value < min_reward) {
-      min_reward = o.value;
-    }
-  }
-  return min_reward;
-}
-
-float max(const joined_event &event) {
-  float max_reward = std::numeric_limits<float>::min();
-  for (const auto &o : event.outcome_events) {
-    if (o.value > max_reward) {
-      max_reward = o.value;
-    }
-  }
-  return max_reward;
-}
-
-float median(const joined_event &event) {
-  std::vector<float> values;
-  for (const auto &o : event.outcome_events) {
-    values.push_back(o.value);
-  }
-
-  int outcome_events_size = values.size();
-
-  sort(values.begin(), values.end());
-  if (outcome_events_size % 2 == 0) {
-    return (values[outcome_events_size / 2 - 1] +
-            values[outcome_events_size / 2]) /
-           2;
-  } else {
-    return values[outcome_events_size / 2];
-  }
-}
-
-float earliest(const joined_event &event) {
-  auto oldest_valid_observation = TimePoint::max();
-  float earliest_reward = 0.f;
-
-  for (const auto &o : event.outcome_events) {
-    if (o.enqueued_time_utc < oldest_valid_observation) {
-      oldest_valid_observation = o.enqueued_time_utc;
-      earliest_reward = o.value;
-    }
-  }
-
-  return earliest_reward;
-}
-} // namespace RewardFunctions
 
 example_joiner::example_joiner(vw *vw)
-    : _vw(vw), _reward_calculation(&RewardFunctions::earliest) {}
+    : _vw(vw), _reward_calculation(&reward::earliest) {}
 
 example_joiner::example_joiner(vw *vw, bool binary_to_json,
                                std::string outfile_name)
-    : _vw(vw), _reward_calculation(&RewardFunctions::earliest),
+    : _vw(vw), _reward_calculation(&reward::earliest),
       _binary_to_json(binary_to_json) {
   _outfile.open(outfile_name, std::ofstream::out);
 }
@@ -191,30 +118,29 @@ void example_joiner::set_problem_type_config(
 }
 
 void example_joiner::set_reward_function(const v2::RewardFunctionType type) {
-  using namespace RewardFunctions;
 
   switch (type) {
   case v2::RewardFunctionType_Earliest:
-    _reward_calculation = &earliest;
+    _reward_calculation = &reward::earliest;
     break;
   case v2::RewardFunctionType_Average:
-    _reward_calculation = &average;
+    _reward_calculation = &reward::average;
     break;
 
   case v2::RewardFunctionType_Sum:
-    _reward_calculation = &sum;
+    _reward_calculation = &reward::sum;
     break;
 
   case v2::RewardFunctionType_Min:
-    _reward_calculation = &min;
+    _reward_calculation = &reward::min;
     break;
 
   case v2::RewardFunctionType_Max:
-    _reward_calculation = &max;
+    _reward_calculation = &reward::max;
     break;
 
   case v2::RewardFunctionType_Median:
-    _reward_calculation = &median;
+    _reward_calculation = &reward::median;
     break;
 
   default:
