+++ external_parser/example_joiner.h
@@ -3,78 +3,14 @@
 #include "err_constants.h"
 
 #include "example.h"
-#include "generated/v2/CbEvent_generated.h"
-#include "generated/v2/FileFormat_generated.h"
 #include "lru_dedup_cache.h"
-#include "timestamp_helper.h"
+#include "reward.h"
 #include "v_array.h"
 
 #include <fstream>
 #include <list>
 #include <queue>
 #include <unordered_map>
-// VW headers
-// vw.h has to come before json_utils.h
-// clang-format off
-#include "vw.h"
-#include "json_utils.h"
-// clang-format on
-
-namespace v2 = reinforcement_learning::messages::flatbuff::v2;
-
-struct metadata_info {
-  TimePoint client_time_utc;
-  std::string app_id;
-  v2::PayloadType payload_type;
-  float pass_probability;
-  v2::EventEncoding event_encoding;
-  std::string event_id;
-  v2::LearningModeType learning_mode;
-};
-
-struct outcome_event {
-  metadata_info metadata;
-  std::string s_index;
-  int index;
-  std::string s_value;
-  float value;
-  TimePoint enqueued_time_utc;
-  bool action_taken;
-};
-
-// TODO in separate PR move metadat_info, outcome_event, joined_event and reward
-// calculations to their own file
-struct joined_event {
-  joined_event(const TimePoint &tp, metadata_info &&mi,
-               DecisionServiceInteraction &&id, const std::string &ctx,
-               const std::string &mid)
-      : joined_event_timestamp(tp), interaction_metadata(std::move(mi)),
-        interaction_data(std::move(id)), outcome_events({}), context(ctx),
-        model_id(mid), ok(true) {}
-  joined_event() : ok(true) {}
-
-  TimePoint joined_event_timestamp;
-  metadata_info interaction_metadata;
-  DecisionServiceInteraction interaction_data;
-  std::vector<outcome_event> outcome_events;
-  std::string context;
-  std::string model_id;
-  bool ok; // ok till proved otherwise
-  // Default Baseline Action for CB is 1 (rl client recommended actions are 1
-  // indexed in the CB case)
-  static const int baseline_action = 1;
-};
-
-using RewardCalcType = float (*)(const joined_event &);
-
-namespace RewardFunctions {
-float average(const joined_event &event);
-float sum(const joined_event &event);
-float min(const joined_event &event);
-float max(const joined_event &event);
-float median(const joined_event &event);
-float apprentice(const joined_event &event);
-} // namespace RewardFunctions
 
 class example_joiner {
 public:
@@ -177,7 +113,7 @@ private:
   float _reward = _default_reward;
   // original reward is used to record the observed reward of apprentice mode
   float _original_reward = _default_reward;
-  RewardCalcType _reward_calculation;
+  reward::RewardFunctionType _reward_calculation;
 
   v2::LearningModeType _learning_mode_config = v2::LearningModeType_Online;
   v2::ProblemType _problem_type_config = v2::ProblemType_UNKNOWN;
